add new method

+++ src/BonusMilesServise.java
@@ -1,5 +1,5 @@
 public class BonusMilesServise {
     public int calculate(int inPrice) {
-        return (int) inPrice / 20;
+        return inPrice / 20;
     }
 }
